Reduce Windows stubs
Commit 421b4cab7d101dabcb4a1ff6348ce9b695431b0b changed the use of
syscall.Signal in favor of unix.Signal, to match other code.

As a result, some code that imported these files on Windows (even if unused)
failed to compile, which lead to d6ba49689be98a4278ce479200f72e1f25404dc4,
which created a separate implementation for Windows without PdeathSignal.

Given that golang.org/x/sys/unix.Signal is an alias for syscall.Signal
(see https://github.com/golang/sys/blob/751e447fb3d0a97f584890476adddc1d56307388/unix/aliases.go#L13-L14),
they should be interchangeable.

Note that the io.go files could probably be excluded on Windows as a whole,
but taking a slightly less rigorous approach in this PR.
